third_party/blink/web_tests/platform/generic/fast/dom/elementsFromPoint/elementsFromPoint-simple-expected.txt

PASS document.elementsFromPoint(1, 1).indexOf(document.getElementById('simpleDiv')) is >= 0
PASS All elements in document.elementsFromPoint(1, 1) intersect (1, 1)
PASS document.elementsFromPoint(199, 1).indexOf(document.getElementById('simpleDiv')) is >= 0
PASS All elements in document.elementsFromPoint(199, 1) intersect (199, 1)
PASS document.elementsFromPoint(1, 199).indexOf(document.getElementById('simpleDiv')) is >= 0
PASS All elements in document.elementsFromPoint(1, 199) intersect (1, 199)
PASS document.elementsFromPoint(199, 199).indexOf(document.getElementById('simpleDiv')) is >= 0
PASS All elements in document.elementsFromPoint(199, 199) intersect (199, 199)
PASS document.elementsFromPoint(51, 51).indexOf(document.getElementById('divWithPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(51, 51) intersect (51, 51)
PASS document.elementsFromPoint(149, 51).indexOf(document.getElementById('divWithPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(149, 51) intersect (149, 51)
PASS document.elementsFromPoint(51, 149).indexOf(document.getElementById('divWithPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(51, 149) intersect (51, 149)
PASS document.elementsFromPoint(149, 149).indexOf(document.getElementById('divWithPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(149, 149) intersect (149, 149)
PASS document.elementsFromPoint(101, 101).indexOf(document.getElementById('divBetweenPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(101, 101) intersect (101, 101)
PASS document.elementsFromPoint(199, 101).indexOf(document.getElementById('divBetweenPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(199, 101) intersect (199, 101)
PASS document.elementsFromPoint(101, 199).indexOf(document.getElementById('divBetweenPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(101, 199) intersect (101, 199)
PASS document.elementsFromPoint(199, 199).indexOf(document.getElementById('divBetweenPseudo')) is >= 0
PASS All elements in document.elementsFromPoint(199, 199) intersect (199, 199)
PASS document.elementsFromPoint(1, 186).indexOf(document.getElementById('withMargins')) is >= 0
PASS All elements in document.elementsFromPoint(1, 186) intersect (1, 186)
PASS document.elementsFromPoint(199, 186).indexOf(document.getElementById('withMargins')) is >= 0
PASS All elements in document.elementsFromPoint(199, 186) intersect (199, 186)
PASS document.elementsFromPoint(1, 384).indexOf(document.getElementById('withMargins')) is >= 0
PASS All elements in document.elementsFromPoint(1, 384) intersect (1, 384)
PASS document.elementsFromPoint(199, 384).indexOf(document.getElementById('withMargins')) is >= 0
PASS All elements in document.elementsFromPoint(199, 384) intersect (199, 384)
PASS document.elementsFromPoint(101, 186).indexOf(document.getElementById('inlineSpan')) is >= 0
PASS All elements in document.elementsFromPoint(101, 186) intersect (101, 186)
PASS document.elementsFromPoint(199, 186).indexOf(document.getElementById('inlineSpan')) is >= 0
PASS All elements in document.elementsFromPoint(199, 186) intersect (199, 186)
PASS document.elementsFromPoint(101, 200).indexOf(document.getElementById('inlineSpan')) is >= 0
PASS All elements in document.elementsFromPoint(101, 200) intersect (101, 200)
PASS document.elementsFromPoint(199, 200).indexOf(document.getElementById('inlineSpan')) is >= 0
PASS All elements in document.elementsFromPoint(199, 200) intersect (199, 200)
PASS document.elementsFromPoint(51, 51).indexOf(document.getElementById('noPointerEvents')) is -1
PASS All elements in document.elementsFromPoint(51, 51) intersect (51, 51)
PASS document.elementsFromPoint(149, 51).indexOf(document.getElementById('noPointerEvents')) is -1
PASS All elements in document.elementsFromPoint(149, 51) intersect (149, 51)
PASS document.elementsFromPoint(51, 349).indexOf(document.getElementById('noPointerEvents')) is -1
PASS All elements in document.elementsFromPoint(51, 349) intersect (51, 349)
PASS document.elementsFromPoint(149, 349).indexOf(document.getElementById('noPointerEvents')) is -1
PASS All elements in document.elementsFromPoint(149, 349) intersect (149, 349)
PASS document.elementsFromPoint(41, 41).indexOf(document.getElementById('threeD')) is >= 0
PASS All elements in document.elementsFromPoint(41, 41) intersect (41, 41)
PASS document.elementsFromPoint(239, 41).indexOf(document.getElementById('threeD')) is >= 0
PASS All elements in document.elementsFromPoint(239, 41) intersect (239, 41)
PASS document.elementsFromPoint(41, 89).indexOf(document.getElementById('threeD')) is >= 0
PASS All elements in document.elementsFromPoint(41, 89) intersect (41, 89)
PASS document.elementsFromPoint(239, 89).indexOf(document.getElementById('threeD')) is >= 0
PASS All elements in document.elementsFromPoint(239, 89) intersect (239, 89)
PASS nodeListToString(document.elementsFromPoint(51, 51)) is "DIV#threeD, DIV#divWithPseudo, DIV#simpleDiv, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(149, 149)) is "DIV#divWithPseudo, DIV#divBetweenPseudo, DIV#divWithPseudo, DIV#simpleDiv, BODY, HTML"
PASS successfullyParsed is true

TEST COMPLETE
